+++ changelog.md
@@ -7,6 +7,11 @@
  - Add snapping to grid. See https://github.com/25A0/Quadtastic/wiki/Using-Quadtastic#grid for more information
  - Fix an issue where files would be listed out of alphabetic order on some
    file systems
+ - Fix the reason why double-clicking an element to rename it only worked for
+   top-level quads and groups ([#29](https://github.com/25A0/Quadtastic/issues/29))
+ - Reserved Lua keywords can now be used as names for quads and groups
+   ([#27](https://github.com/25A0/Quadtastic/issues/27))
+ - The wiki has moved to https://github.com/25A0/Quadtastic/blob/master/wiki ([#26](https://github.com/25A0/Quadtastic/issues/26))
 
 ### Release 0.6.2, 2017-08-23